+++ react-stage/src/App.js
@@ -22,7 +22,9 @@ import SuperviseurEtudiantsAssignation from './component/Superviseur/Superviseur
 import Dashboard from './component/dashboard/Dashboard';
 import Contrat from './component/contrat/Contrat';
 import AllSessionInfo from "./component/allSessionViewer/AllSessionInfo";
-import NewContrat from "./component/contrat/newContrat/NewContrat"
+import NewContrat from "./component/contrat/newContrat/NewContratForm"
+import NewContratForm from "./component/contrat/newContrat/NewContratForm";
+import NewContratSuccess from "./component/contrat/newContrat/NewContratSuccess";
 
 
 function App() {
@@ -49,6 +51,9 @@ function App() {
               <Route exact path="/gestion/superviseur"><SuperviseurEtudiantsAssignation /></Route>
               <Route exact path="/gestion/allSession"><AllSessionInfo /></Route>
               <Route exact path="/gestion/newContrat"><NewContrat /></Route>
+              <Route exact path="/gestion/newContratForm"><NewContratForm /></Route>
+              <Route exact path="/gestion/newContratSucces"><NewContratSuccess /></Route>
+
             </Switch>
           </div>
         </UserInfo>
